ansachs/requirements.txt
boto3==1.18.23
python-dotenv==0.19.0
